MidiClockToCV: big update.  Pots and Pushbuttons.
Big update to match the newly-revised hardware (I finally added the pots
and pushbuttons to the Sparkfun MIDI Shield).  The two output channels
are now fully independent and the two pots set the MIDI divider for each
channel.  The middle pushbutton resets the timing, though it seems to
lag.  The pushbuttons and pots need more work to feel right.